test(agent-status): add component tests and update integration tests
