+++ console/src/AgentStatus/Component.tsx
@@ -1,0 +1,465 @@
+import { Container } from "automated-gameplay-transmitter";
+import type { ReactNode } from "react";
+import { useState } from "react";
+import type { AgentStatusRow } from "./sections/AgentStatusSectionCard";
+import { GAME_SECTION_TITLE, GameStatusSection } from "./sections/GameStatusSection";
+import { LIVE_DELIVERY_SECTION_TITLE, LiveDeliveryStatusSection } from "./sections/LiveDeliveryStatusSection";
+import { MARKOV_MODEL_SECTION_TITLE, MarkovModelStatusSection } from "./sections/MarkovModelStatusSection";
+
+export type AgentStateResponse = {
+    niconama?: any;
+    currentGame?: { name?: string; state?: any } | null;
+    nGram?: number;
+    nGramRaw?: number;
+    speechHistory?: Array<{ id?: string; speech?: string; nGram?: number; nGramRaw?: number }>;
+    speech?: { speech?: string; silent?: boolean };
+    canSpeak?: boolean;
+};
+
+export const parseAgentStateResponse = (responseText: string): AgentStateResponse => {
+    try {
+        return JSON.parse(responseText) as AgentStateResponse;
+    } catch (err) {
+        throw new Error("配信状態の応答形式が不正です。");
+    }
+};
+
+export const createAgentStateWebSocketUrl = (baseHref: `wss:${string}`) => {
+    const normalized = baseHref.endsWith("/") ? baseHref.slice(0, -1) : baseHref;
+    return `${normalized}/api/ws`;
+};
+
+export const AGENT_STATE_REFRESH_INTERVAL_MS = 1_000;
+export const AGENT_STATE_MOCK_NOTICE_MESSAGE = "配信エージェント状態モックを表示中";
+const SPEECH_UNAVAILABLE_INDICATOR = "・・・";
+
+// Distinguishes unix seconds from unix milliseconds by treating 13-digit values as milliseconds.
+const UNIX_MILLISECONDS_THRESHOLD = 1_000_000_000_000;
+const LIVE_DELIVERY_ROW_LABELS = ["配信指標", "タイトル", "配信URL", "開始時刻", "発話内容"] as const;
+const MARKOV_MODEL_ROW_LABELS = ["生成N-gram", "これまでの発話"] as const;
+const GAME_ROW_LABELS = ["現在のゲーム", "ゲーム情報"] as const;
+const GAME_STATE_EMPTY_ARRAY_LABEL = "(空の配列)";
+const GAME_STATE_EMPTY_OBJECT_LABEL = "(空のオブジェクト)";
+// 上から順に: 上部固定領域（見出し・更新時刻・通知類）/ 詳細一覧（残り高さをすべて使用）
+const AGENT_STATUS_GRID_ROW_TEMPLATE_CLASS = "grid-rows-[auto_minmax(0,1fr)]";
+const createLabelSet = (labels: readonly string[]) => new Set<string>(labels);
+const LIVE_DELIVERY_ROW_LABEL_SET = createLabelSet(LIVE_DELIVERY_ROW_LABELS);
+const MARKOV_MODEL_ROW_LABEL_SET = createLabelSet(MARKOV_MODEL_ROW_LABELS);
+const GAME_ROW_LABEL_SET = createLabelSet(GAME_ROW_LABELS);
+
+const formatStateLabel = (type: string | undefined): string => {
+    if (type === "live") {
+        return "配信中";
+    }
+    if (type === "offline") {
+        return "停止中";
+    }
+    return type ?? "-";
+};
+
+const formatStartDate = (startAtUnixTime: number | undefined): string => {
+    if (typeof startAtUnixTime !== "number" || !Number.isFinite(startAtUnixTime) || startAtUnixTime <= 0) {
+        return "-";
+    }
+    const startAtUnixTimeMilliseconds = startAtUnixTime >= UNIX_MILLISECONDS_THRESHOLD
+        ? startAtUnixTime
+        : startAtUnixTime * 1000;
+    return new Date(startAtUnixTimeMilliseconds).toLocaleString("ja-JP");
+};
+
+const formatMetricValue = (metricValue: number | undefined): string => {
+    return metricValue === undefined ? "-" : String(metricValue);
+};
+
+const formatNGramValue = (nGram: number | undefined, nGramRaw: number | undefined): string => {
+    if (nGram === undefined || !Number.isFinite(nGram) || nGram < 1) {
+        return "-";
+    }
+    const nGramValue = `${Math.floor(nGram)}-gram`;
+    if (nGramRaw === undefined || !Number.isFinite(nGramRaw) || nGramRaw < 1) {
+        return nGramValue;
+    }
+    const formattedRaw = Number(nGramRaw).toFixed(2);
+    return `${nGramValue} (${formattedRaw})`;
+};
+
+const formatSpeechHistoryItemText = (
+    speechText: string,
+    nGram: number | undefined,
+): string => {
+    return `${speechText} (${formatSpeechHistoryNGramLabel(nGram)})`;
+};
+
+const formatSpeechHistoryNGramLabel = (nGram: number | undefined): string => {
+    if (nGram === undefined || !Number.isFinite(nGram) || nGram < 1) {
+        return "-";
+    }
+    return `${Math.floor(nGram)}g`;
+};
+
+const createSpeechHistoryDisplayItems = (
+    speechHistory: AgentStateResponse["speechHistory"] | undefined,
+): Array<{ id: string; speechText: string; displayLine: string; nGramLabel: string }> => {
+    if (!Array.isArray(speechHistory)) {
+        return [];
+    }
+    const speechHistoryItems = speechHistory.reduce<Array<{ id: string; speechText: string; displayLine: string; nGramLabel: string }>>(
+        (accumulatedItems, speechHistoryItem) => {
+            const speechText = speechHistoryItem.speech?.trim();
+            if (!speechText) {
+                return accumulatedItems;
+            }
+            const displayOrder = accumulatedItems.length + 1;
+            accumulatedItems.push({
+                id: speechHistoryItem.id ?? `speech-history-${displayOrder}`,
+                speechText,
+                displayLine: formatSpeechHistoryItemText(speechText, speechHistoryItem.nGram),
+                nGramLabel: formatSpeechHistoryNGramLabel(speechHistoryItem.nGram),
+            });
+            return accumulatedItems;
+        },
+        [],
+    );
+
+    const itemsWithSeq = speechHistoryItems.map((item) => {
+        const match = item.id.match(/(\d+)$/);
+        return { ...item, seq: match ? Number(match[1]) : undefined };
+    });
+
+    const seqCount = itemsWithSeq.reduce((c, it) => (it.seq !== undefined ? c + 1 : c), 0);
+    if (seqCount >= 2) {
+        itemsWithSeq.sort((a, b) => (b.seq ?? 0) - (a.seq ?? 0));
+        return itemsWithSeq.map(({ seq, ...rest }) => rest);
+    }
+
+    return itemsWithSeq.map(({ seq, ...rest }) => rest);
+};
+
+const createSpeechHistoryValueComponent = (
+    speechHistory: AgentStateResponse["speechHistory"] | undefined,
+): ReactNode => {
+    const speechHistoryItems = createSpeechHistoryDisplayItems(speechHistory);
+    if (speechHistoryItems.length === 0) {
+        return <span>-</span>;
+    }
+    return (
+        <ul className="grid grid-cols-1 gap-2">
+            {speechHistoryItems.map((speechHistoryItem) => (
+                <li key={speechHistoryItem.id} className="rounded-md border border-emerald-300/30 p-2">
+                    <div className="grid grid-cols-[minmax(0,1fr)_auto_auto] items-start gap-2">
+                        <p className="truncate">{speechHistoryItem.speechText}</p>
+                        <span className="text-xs text-emerald-200 whitespace-nowrap">{speechHistoryItem.nGramLabel}</span>
+                        <button
+                            type="button"
+                            disabled
+                            aria-label="学習の取り消し"
+                            title="学習の取り消し"
+                            className="text-base leading-none h-7 w-7 rounded border border-emerald-300/50 text-emerald-200 opacity-70 cursor-not-allowed flex items-center justify-center"
+                        >
+                            ↩
+                        </button>
+                    </div>
+                </li>
+            ))}
+        </ul>
+    );
+};
+
+const createLiveDeliveryMetricsValueComponent = (niconamaState: AgentStateResponse["niconama"]): ReactNode => {
+    const liveMetricItems = [
+        { label: "状態", value: formatStateLabel(niconamaState?.type) },
+        { label: "視聴者数", value: formatMetricValue(niconamaState?.meta?.total?.listeners) },
+        { label: "コメント数", value: formatMetricValue(niconamaState?.meta?.total?.comments) },
+        { label: "ギフト", value: formatMetricValue(niconamaState?.meta?.total?.gift) },
+        { label: "広告", value: formatMetricValue(niconamaState?.meta?.total?.ad) },
+    ];
+    return (
+        <div className="rounded-md border border-emerald-300/30 p-2">
+            <div className="grid grid-cols-5 gap-x-2 gap-y-1">
+                {liveMetricItems.map((liveMetricItem) => (
+                    <p key={liveMetricItem.label} className="font-bold text-center whitespace-nowrap">{liveMetricItem.label}</p>
+                ))}
+                {liveMetricItems.map((liveMetricItem) => (
+                    <p key={`${liveMetricItem.label}-value`} className="text-center whitespace-nowrap">{liveMetricItem.value}</p>
+                ))}
+            </div>
+        </div>
+    );
+};
+
+const formatCurrentGameStateLeafValue = (stateValue: unknown): string => {
+    if (stateValue === null) {
+        return "null";
+    }
+    if (stateValue === undefined) {
+        return "-";
+    }
+    return String(stateValue);
+};
+
+const renderCurrentGameStateValueComponent = (
+    currentGameStateValue: unknown,
+    visitedObjects = new WeakSet<object>(),
+): ReactNode => {
+    if (currentGameStateValue === null || typeof currentGameStateValue !== "object") {
+        return <span>{formatCurrentGameStateLeafValue(currentGameStateValue)}</span>;
+    }
+
+    if (visitedObjects.has(currentGameStateValue as object)) {
+        throw new TypeError("Cannot render currentGame.state because circular references were detected (will display '-').");
+    }
+
+    visitedObjects.add(currentGameStateValue as object);
+    try {
+        if (Array.isArray(currentGameStateValue)) {
+            if (currentGameStateValue.length === 0) {
+                return <span>{GAME_STATE_EMPTY_ARRAY_LABEL}</span>;
+            }
+            return (
+                <ul className="list-disc pl-5 space-y-1">
+                    {currentGameStateValue.map((arrayItem, arrayIndex) => {
+                        const arrayItemKey =
+                            arrayItem !== null && typeof arrayItem === "object"
+                                ? `array-item-object-${arrayIndex}`
+                                : `array-item-${formatCurrentGameStateLeafValue(arrayItem)}-${arrayIndex}`;
+                        return <li key={arrayItemKey}>{renderCurrentGameStateValueComponent(arrayItem, visitedObjects)}</li>;
+                    })}
+                </ul>
+            );
+        }
+
+        const objectEntries = Object.entries(currentGameStateValue as Record<string, unknown>);
+        if (objectEntries.length === 0) {
+            return <span>{GAME_STATE_EMPTY_OBJECT_LABEL}</span>;
+        }
+        return (
+            <ul className="space-y-1">
+                {objectEntries.map(([stateKey, stateValue]) => {
+                    if (stateValue !== null && typeof stateValue === "object") {
+                        return (
+                            <li key={stateKey}>
+                                <div className="font-semibold">{stateKey}</div>
+                                <div className="pl-4 border-l border-emerald-300/40">
+                                    {renderCurrentGameStateValueComponent(stateValue, visitedObjects)}
+                                </div>
+                            </li>
+                        );
+                    }
+                    return (
+                        <li key={stateKey}>
+                            <span className="font-semibold">{stateKey}</span>
+                            <span>: {formatCurrentGameStateLeafValue(stateValue)}</span>
+                        </li>
+                    );
+                })}
+            </ul>
+        );
+    } finally {
+        visitedObjects.delete(currentGameStateValue as object);
+    }
+};
+
+const createCurrentGameInfoValueComponent = (currentGameState: Record<string, unknown> | undefined): ReactNode => {
+    if (currentGameState === undefined) {
+        return <span>-</span>;
+    }
+
+    try {
+        return renderCurrentGameStateValueComponent(currentGameState);
+    } catch {
+        return <span>-</span>;
+    }
+};
+
+export const createAgentStatusRows = (stateResponse: AgentStateResponse | null): AgentStatusRow[] => {
+    const rows: AgentStatusRow[] = [];
+
+    const niconamaState = stateResponse?.niconama;
+    if (niconamaState && Object.keys(niconamaState).length > 0) {
+        rows.push(
+            { label: "配信指標", hideLabel: true, valueComponent: createLiveDeliveryMetricsValueComponent(niconamaState) },
+            { label: "タイトル", value: niconamaState.meta?.title ?? "-" },
+            { label: "配信URL", value: niconamaState.meta?.url ?? "-", href: niconamaState.meta?.url },
+            { label: "開始時刻", value: formatStartDate(niconamaState.meta?.start) },
+        );
+    }
+
+    if (stateResponse !== null && stateResponse !== undefined && "currentGame" in stateResponse) {
+        rows.push({ label: "現在のゲーム", value: stateResponse.currentGame?.name ?? "-" });
+        rows.push({
+            label: "ゲーム情報",
+            valueComponent: createCurrentGameInfoValueComponent(stateResponse.currentGame?.state),
+        });
+    }
+
+    if (stateResponse?.nGram !== undefined) {
+        rows.push({ label: "生成N-gram", value: formatNGramValue(stateResponse.nGram, stateResponse.nGramRaw) });
+    }
+
+    if (createSpeechHistoryDisplayItems(stateResponse?.speechHistory).length > 0) {
+        rows.push({
+            label: "これまでの発話",
+            valueComponent: createSpeechHistoryValueComponent(stateResponse?.speechHistory),
+        });
+    }
+
+    if (stateResponse?.canSpeak === false) {
+        rows.push({ label: "発話内容", value: SPEECH_UNAVAILABLE_INDICATOR });
+    } else if (stateResponse?.speech !== undefined) {
+        rows.push({ label: "発話内容", value: stateResponse.speech.speech ?? "-" });
+    }
+
+    return rows;
+};
+
+export const createAgentStatusSections = (stateResponse: AgentStateResponse | null): { title: string; rows: AgentStatusRow[] }[] => {
+    const rows = createAgentStatusRows(stateResponse);
+    const liveDeliveryRows = rows.filter((row) => LIVE_DELIVERY_ROW_LABEL_SET.has(row.label));
+    const markovModelRows = rows.filter((row) => MARKOV_MODEL_ROW_LABEL_SET.has(row.label));
+    const gameRows = rows.filter((row) => GAME_ROW_LABEL_SET.has(row.label));
+
+    const sections = [
+        { title: LIVE_DELIVERY_SECTION_TITLE, rows: liveDeliveryRows },
+        { title: MARKOV_MODEL_SECTION_TITLE, rows: markovModelRows },
+        { title: GAME_SECTION_TITLE, rows: gameRows },
+    ];
+
+    return sections.filter((section) => section.rows.length > 0);
+};
+
+export type AgentStatusViewProps = {
+    agentStateResponse: AgentStateResponse | null;
+    agentStatusError: string | null;
+    lastUpdatedTime: string;
+    isLoadingAgentState: boolean;
+    isShowingMockAgentState: boolean;
+    onRefresh?: () => void | Promise<void>;
+};
+
+export function AgentStatusView({
+    agentStateResponse,
+    agentStatusError,
+    lastUpdatedTime,
+    isLoadingAgentState,
+    isShowingMockAgentState,
+    onRefresh,
+}: AgentStatusViewProps) {
+    const agentStatusSections = createAgentStatusSections(agentStateResponse);
+    const sectionMap = agentStatusSections.reduce<Partial<Record<string, { title: string; rows: AgentStatusRow[] }>>>(
+        (accumulatedSections, section) => {
+            accumulatedSections[section.title] = section;
+            return accumulatedSections;
+        },
+        {},
+    );
+    const liveDeliverySection = sectionMap[LIVE_DELIVERY_SECTION_TITLE];
+    const markovModelSection = sectionMap[MARKOV_MODEL_SECTION_TITLE];
+    const gameSection = sectionMap[GAME_SECTION_TITLE];
+    const hasPrimaryColumnSections = liveDeliverySection !== undefined || gameSection !== undefined;
+
+    return (
+        <div className={`mx-auto w-full max-w-7xl h-full min-h-0 text-left grid ${AGENT_STATUS_GRID_ROW_TEMPLATE_CLASS} gap-4`}>
+            <div className="flex flex-col gap-4">
+                <div className="flex flex-wrap items-center justify-between gap-3">
+                    <h1 className="text-2xl font-bold">
+                        <a href="https://live.nicovideo.jp/watch/user/14171889" target="_blank" rel="noopener noreferrer">
+                            馬可無序
+                        </a>
+                    </h1>
+                    <p className="text-sm text-emerald-200 whitespace-nowrap">
+                        最終更新: {lastUpdatedTime || "未取得"}
+                    </p>
+                    <button
+                        type="button"
+                        onClick={() => {
+                            if (onRefresh) void onRefresh();
+                        }}
+                        disabled={isLoadingAgentState}
+                        className="bg-emerald-300 text-emerald-950 border-0 px-5 py-1.5 rounded-lg font-bold transition-all duration-100 hover:bg-emerald-200 hover:-translate-y-px cursor-pointer whitespace-nowrap disabled:opacity-60 disabled:cursor-not-allowed"
+                    >
+                        更新
+                    </button>
+                </div>
+                {isShowingMockAgentState ? (
+                    <Container>
+                        <div
+                            data-testid="agent-status-mock-notice"
+                            className="w-full bg-emerald-950/70 border-2 border-emerald-300 rounded-xl p-3 text-emerald-50"
+                        >
+                            {AGENT_STATE_MOCK_NOTICE_MESSAGE}
+                        </div>
+                    </Container>
+                ) : null}
+                {agentStatusError ? (
+                    <div
+                        data-testid="agent-status-error"
+                        className="w-full min-h-[80px] bg-red-950/60 border-2 border-red-300 rounded-xl p-3 text-red-100"
+                    >
+                        取得エラー: {agentStatusError}
+                    </div>
+                ) : null}
+            </div>
+            {agentStatusSections.length === 0 ? (
+                <Container>
+                    <div
+                        data-testid="agent-status-empty"
+                        className="w-full min-h-[80px] bg-emerald-950/70 border-2 border-emerald-300 rounded-xl p-3 text-emerald-50"
+                    >
+                        {isLoadingAgentState ? "読み込み中..." : "配信情報はありません。"}
+                    </div>
+                </Container>
+            ) : (
+                <div
+                    data-testid="agent-status-details"
+                    className="w-full h-full min-h-0 overflow-y-auto pr-1 grid grid-cols-1 xl:grid-cols-[minmax(0,1fr)_minmax(0,2fr)] auto-rows-min xl:auto-rows-[minmax(0,1fr)] gap-4"
+                >
+                    {hasPrimaryColumnSections ? (
+                        <div className="min-w-0 min-h-0 h-full flex flex-col gap-4">
+                            {liveDeliverySection ? <LiveDeliveryStatusSection liveDeliveryRows={liveDeliverySection.rows} /> : null}
+                            {gameSection ? <GameStatusSection gameRows={gameSection.rows} /> : null}
+                        </div>
+                    ) : null}
+                    {markovModelSection ? (
+                        <div className={`min-w-0 min-h-0 h-full${hasPrimaryColumnSections ? " xl:col-start-2" : ""}`}>
+                            <MarkovModelStatusSection markovModelRows={markovModelSection.rows} />
+                        </div>
+                    ) : null}
+                </div>
+            )}
+        </div>
+    );
+}
+
+export function AgentStatus() {
+    const [state, setState] = useState(() => ({
+        agentStateResponse: null as AgentStateResponse | null,
+        agentStatusError: null as string | null,
+        lastUpdatedTime: "",
+        isLoadingAgentState: false,
+        isShowingMockAgentState: false,
+    }));
+
+    const refresh = () => {
+        // noop: networking is managed elsewhere; this keeps the component usable in tests
+        return;
+    };
+
+    const {
+        agentStateResponse,
+        agentStatusError,
+        lastUpdatedTime,
+        isLoadingAgentState,
+        isShowingMockAgentState,
+    } = state;
+
+    return (
+        <AgentStatusView
+            agentStateResponse={agentStateResponse}
+            agentStatusError={agentStatusError}
+            lastUpdatedTime={lastUpdatedTime}
+            isLoadingAgentState={isLoadingAgentState}
+            isShowingMockAgentState={isShowingMockAgentState}
+            onRefresh={refresh}
+        />
+    );
+}
